requirements.txt
requests>=2.28.2
python-dotenv>=0.21.1
redis>=4.5.1
flask>=2.2.3